+++ muna/beta/metadata/vllm.py
@@ -8,7 +8,10 @@ from typing import Annotated, Literal
 
 from ._torch import _validate_torch_module
 
-class vLLMInferenceMetadata(BaseModel, **ConfigDict(arbitrary_types_allowed=True, frozen=True)):
+class TorchTovLLMInferenceMetadata(
+    BaseModel,
+    **ConfigDict(arbitrary_types_allowed=True, frozen=True)
+):
     """
     Metadata to compile a large language model for multi-GPU inference with vLLM.